refactor(core): name admin-bar catalogs per surface
Rename core's locale catalogs from the generic locales/{locale}.po to
locales/admin-bar-{locale}.po so a later debug-bar catalog can sit
beside them in one flat dir without colliding. defineLinguiConfig
gains catalogPath + include options for the per-surface path; the
admin bar imports the renamed compiled catalogs. resolveBarLocale
keeps reading user.meta.locale and now documents why it must stay
separate from the path-gated unified resolveLocale: the bar localizes
per authenticated user on public routes, which is cache-safe only
because session-bearing responses are private anyway, and the locale
cookie (Path=/_plumix/) never reaches the front end.
