+++ packages/core/locales/admin-bar-ar.d.mts
