+++ packages/core/locales/admin-bar-ar.mjs
