+++ packages/core/locales/admin-bar-ar.po
