+++ packages/core/locales/admin-bar-de.d.mts
