+++ packages/core/locales/admin-bar-de.mjs
