+++ packages/core/locales/admin-bar-de.po
