+++ packages/core/locales/admin-bar-en.d.mts
