+++ packages/core/locales/admin-bar-en.mjs
