+++ packages/core/locales/admin-bar-uk.d.mts
