+++ packages/core/locales/admin-bar-uk.mjs
