+++ packages/core/locales/admin-bar-uk.po
