+++ packages/core/locales/admin-bar-zh-CN.d.mts
