+++ packages/core/locales/admin-bar-zh-CN.po
