+++ packages/core/src/admin-bar/i18n.ts
@@ -1,15 +1,17 @@
-// Admin-bar strings live in `locales/*.po` like every other plumix
-// surface — `plumix i18n verify` gates the descriptor↔catalog drift
-// and translators never have to find a second system. Catalogs are
-// compiled to static modules (worker-safe, no fs) and imported via
-// the package's own `./locales/*` subpath; the SSR render does a
-// direct per-request lookup, never a shared-singleton `activate()`.
+// Admin-bar strings live in `locales/admin-bar-*.po` like every other
+// plumix surface — `plumix i18n verify` gates the descriptor↔catalog
+// drift and translators never have to find a second system. Catalogs
+// compile to static modules (worker-safe, no fs) imported via the
+// package's own `./locales/*` subpath; the SSR render does a direct
+// per-request lookup, never a shared-singleton `activate()`. The
+// `admin-bar-` prefix keeps this surface's catalog distinct from a
+// later debug-bar catalog in the same flat locales/ dir.
 
-import { messages as arMessages } from "@plumix/core/locales/ar";
-import { messages as deMessages } from "@plumix/core/locales/de";
-import { messages as enMessages } from "@plumix/core/locales/en";
-import { messages as ukMessages } from "@plumix/core/locales/uk";
-import { messages as zhCnMessages } from "@plumix/core/locales/zh-CN";
+import { messages as arMessages } from "@plumix/core/locales/admin-bar-ar";
+import { messages as deMessages } from "@plumix/core/locales/admin-bar-de";
+import { messages as enMessages } from "@plumix/core/locales/admin-bar-en";
+import { messages as ukMessages } from "@plumix/core/locales/admin-bar-uk";
+import { messages as zhCnMessages } from "@plumix/core/locales/admin-bar-zh-CN";
 
 export type BarLocale = "en" | "de" | "uk" | "ar" | "zh-CN";
 
@@ -50,6 +52,18 @@ const KNOWN: ReadonlySet<string> = new Set(Object.keys(CATALOGS));
  * Resolves the admin user's bar locale from `meta.locale` (per WP semantics —
  * `get_user_locale()`, not `get_locale()`). Falls back to English when the
  * user has no stored locale or stored a locale we don't ship strings for.
+ *
+ * Deliberately NOT the unified `resolveLocale` (`i18n/resolve-locale.ts`):
+ * that resolver path-gates `meta.locale`/Accept-Language/cookie to
+ * `/_plumix/*` so public HTML stays byte-identical per URL for the CDN
+ * cache. The admin bar is the documented exception — it must localize
+ * per authenticated user *on public routes*, and that's cache-safe
+ * precisely because the bar only renders when a session cookie is
+ * present, and session-bearing responses are private/uncacheable
+ * anyway. Anonymous (cacheable) visitors get no bar and the default
+ * locale. The locale cookie can't help here regardless: it's
+ * `Path=/_plumix/`, so the browser never sends it on the front end.
+ * Fold this into `resolveLocale` and per-user bar localization breaks.
  */
 export function resolveBarLocale(user: {
   readonly meta: Record<string, unknown>;
